XD-2671 XD deployments via ZK distributed queue
 - Add ZK distributed queue that holds the deployment requests
   - Create `DeploymentMessage` that wraps the details of the
deployment requests
   - Create `DeploymentMessageProducer` and make it available for
the REST controllers so that the deployment requests get converted
as `DeploymentMessage`s and pushed to the distributed queue by
the producer
   - Create `DeploymentMessageConsumer` which is set only when
the DeploymentSupervisor is elected as leader
   - Introduce `before*check*` methods in `ResourcePreDeployer`
      - These methods are used by the DeploymentMessageProducer
to check the validity of the request before pushing the messages
   - Add `DeploymentHandler` that is responsible for deploy/un-deploy
the deployment units(Stream/Job) by a given name.
   - Have both Stream/Job based deployment handlers
   - The DeploymentMessageConsumer delegates the deployment
requests to these handlers via Stream/JobDeployer.

   - Refactor DeploymentSupervisor
      - Add utility class `ZKDeploymentUtility` that holds the
reference to all the singletons that are used by ZK based
deployments.
   - Move spring-xd-dirt server package classes into `admin`,
`container` and `admin.deployment` packages.